requirements.txt
Pillow
numpy
scikit-learn
matplotlib
openai  # For AI image generation (e.g., using DALL-E or similar models)